+++ harbaughsim16/PlayerF7.h
@@ -9,10 +9,10 @@
 #import "Player.h"
 
 @interface PlayerF7 : Player
-@property (nonatomic) NSInteger ratF7Pow;
-@property (nonatomic) NSInteger ratF7Rsh;
-@property (nonatomic) NSInteger ratF7Pas;
+@property (nonatomic) int ratF7Pow;
+@property (nonatomic) int ratF7Rsh;
+@property (nonatomic) int ratF7Pas;
 
-+(instancetype)newF7WithName:(NSString *)nm team:(Team *)t year:(NSInteger)yr potential:(NSInteger)pot footballIQ:(NSInteger)iq power:(NSInteger)pow rush:(NSInteger)rsh pass:(NSInteger)pass;
-+(instancetype)newF7WithName:(NSString*)nm year:(NSInteger)yr stars:(NSInteger)stars team:(Team*)t;
++(instancetype)newF7WithName:(NSString *)nm team:(Team *)t year:(int)yr potential:(int)pot footballIQ:(int)iq power:(int)pow rush:(int)rsh pass:(int)pass;
++(instancetype)newF7WithName:(NSString*)nm year:(int)yr stars:(int)stars team:(Team*)t;
 @end
