+++ harbaughsim16/PlayerOL.h
@@ -9,11 +9,11 @@
 #import "Player.h"
 
 @interface PlayerOL : Player
-@property (nonatomic) NSInteger ratOLPow;
+@property (nonatomic) int ratOLPow;
 //OLBkR affects how well he blocks for running plays
-@property (nonatomic) NSInteger ratOLBkR;
+@property (nonatomic) int ratOLBkR;
 //OLBkP affects how well he blocks for passing plays
-@property (nonatomic) NSInteger ratOLBkP;
-+(instancetype)newOLWithName:(NSString *)nm team:(Team *)t year:(NSInteger)yr potential:(NSInteger)pot footballIQ:(NSInteger)iq power:(NSInteger)pow rush:(NSInteger)rsh pass:(NSInteger)pass;
-+(instancetype)newOLWithName:(NSString*)nm year:(NSInteger)yr stars:(NSInteger)stars team:(Team*)t;
+@property (nonatomic) int ratOLBkP;
++(instancetype)newOLWithName:(NSString *)nm team:(Team *)t year:(int)yr potential:(int)pot footballIQ:(int)iq power:(int)pow rush:(int)rsh pass:(int)pass;
++(instancetype)newOLWithName:(NSString*)nm year:(int)yr stars:(int)stars team:(Team*)t;
 @end
