+++ harbaughsim16/TeamStrategy.h
@@ -10,15 +10,15 @@
 
 @interface TeamStrategy : NSObject
 
-@property (nonatomic) NSInteger rushYdBonus;
-@property (nonatomic) NSInteger rushAgBonus;
-@property (nonatomic) NSInteger passYdBonus;
-@property (nonatomic) NSInteger passAgBonus;
+@property (nonatomic) int rushYdBonus;
+@property (nonatomic) int rushAgBonus;
+@property (nonatomic) int passYdBonus;
+@property (nonatomic) int passAgBonus;
 
 @property (strong, nonatomic) NSString * stratName;
 @property (strong, nonatomic) NSString * stratDescription;
 
-+(instancetype)newStrategyWithName:(NSString*)name description:(NSString*)description rYB:(NSInteger)rYB rAB:(NSInteger)rAB pYB:(NSInteger)pYB pAB:(NSInteger)pAB;
++(instancetype)newStrategyWithName:(NSString*)name description:(NSString*)description rYB:(int)rYB rAB:(int)rAB pYB:(int)pYB pAB:(int)pAB;
 +(instancetype)newStrategy;
 
 @end
